# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": "Documentação da API do backend"
  },
  "servers": [
    {
      "url": "https://tematicoapi.zapchatbr.com/",
      "description": "API Temático"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/eventos": {
      "get": {
        "summary": "Listar eventos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "data",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "post": {
        "summary": "Criar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "data": {
                    "type": "string",
                    "format": "date"
                  },
                  "hora": {
                    "type": "string",
                    "pattern": "^\\d{2}:\\d{2}$"
                  },
                  "restauranteId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "nome",
                  "data",
                  "hora",
                  "restauranteId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento criado"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
    "/eventos/disponiveis": {
      "get": {
        "summary": "Listar eventos disponíveis por data",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "data",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos disponíveis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "disponibilidade": { "type": "boolean" },
                      "restaurante": { "type": "string" },
                      "vagas": { "type": "integer" },
                      "id": { "type": "integer" },
                      "nome": { "type": "string" },
                      "hora": {
                        "type": "string",
                        "pattern": "^\\d{2}:\\d{2}$"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Data inválida" }
        },
        "tags": [ "Eventos" ]
      }
    },
    "/eventos/{id}": {
      "get": {
        "summary": "Obter evento pelo ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "put": {
        "summary": "Atualizar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "data": {
                    "type": "string",
                    "format": "date"
                  },
                  "hora": {
                    "type": "string"
                  },
                  "restauranteId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento atualizado"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "delete": {
        "summary": "Deletar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento deletado"
          },
          "400": {
            "description": "Evento possui reservas ativas"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
    "/eventos/{id}/disponibilidade": {
      "get": {
        "summary": "Verificar disponibilidade do evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disponibilidade retornada"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
    "/eventos/{eventoId}/marcacoes/{reservaId}/status": {
      "patch": {
        "summary": "Atualizar status da marcação do evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado"
          },
          "404": {
            "description": "Marcação não encontrada"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
...
  }
}